requirements.txt
Django==4.2.7
djangorestframework==3.14.0
Pillow==10.1.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0